add security variable session sneaker_db_id

+++ app/controllers/users/registrations_controller.rb
@@ -38,7 +38,9 @@ class Users::RegistrationsController < Devise::RegistrationsController
           }
         }])
       end
-      session.delete(:sneaker_session_id)
+      unless session[:sneaker_session_id].nil?
+        session.delete(:sneaker_session_id)
+      end
     end
   end
